Refactor to project based profiles
The file loader system is now project based vs. individual file based. Rather than requesting an individual gerber file to be loaded and a profile that specifies how to expose that gerber file, the file loader is given a project id, where a "project" encompasses one or more gerber files and drill files. The UI asks the user to select a project, an action to take (mill, drill, expose), what side to use for the action, and what stock will be used as the action is being processed. All of that information is placed into a complete profile, which is used not only by the file loader system, but the exposure and milling subsystems as well.

Mixed in with this refactor is support for "popup windows" in the UI so error messages and other modal instructions can be presented to the user.